+++ src/main/frontend/views/userPortfolio/UserPortfolioView.tsx
@@ -19,6 +19,7 @@ export default function UserPortfolioView() {
             // Call the getPortfolio function with pan as path parameter and asOfDate as query parameter
             const response = await getPortfolio(pan, asOfDate ?? undefined);
             setPortfolio(response || null);
+            setError(null);
         } catch (e) {
             setError(`Failed to fetch portfolio: ${(e as Error).message}`);
         }
