+++ src/main/java/com/app/folioman/shared/MFNavService.java
@@ -7,5 +7,10 @@ public interface MFNavService {
 
     MFSchemeDTO getNavByDateWithRetry(Long schemeId, LocalDate asOfDate);
 
+    /**
+     * Retrieves a list of scheme IDs for which historical data has not been loaded.
+     *
+     * @return A List of Long values representing the scheme IDs without loaded historical data.
+     */
     List<Long> getHistoricalDataNotLoadedSchemeIdList();
 }
